Address Copilot review: GDI+ catch arms, Path.Join, test hygiene
Two suppressed-comment findings (they had no thread, so they are only
visible in the review body):

- ImageProcessor.FrameHasContent and CaptureAsync's per-screenshot filter
  caught ArgumentException but not ExternalException/OutOfMemoryException,
  the other two faces GDI+ shows for a bad image. DiagramProcessor already
  caught all three, so this was an inconsistency inside this PR.

  Honest scope: I could not construct an input that reaches the new arms
  through this path. Measured on this runtime, every corrupt-PNG case
  (magic-only, magic+garbage, truncated, JPEG magic) and the indexed-pixel-
  format case all raise ArgumentException, and ProcessAndWrite writes via
  File.WriteAllBytes, so an unwritable target raises IOException. Bitmap.Save
  to a *path* does raise ExternalException ("A generic error occurred in
  GDI+") for a locked file or a directory, but this code never saves to a
  path. So the widening is defence-in-depth and consistency, not a fix for a
  demonstrated reachable bug — and deliberately without a regression test,
  because any test I could write here would pass identically with the old
  narrow filter and would therefore prove nothing.

Code-quality findings, all valid:

- Path.Combine -> Path.Join at both sites. Combine discards everything
  before a rooted segment, which would let page- or manifest-derived text
  produce an absolute path *before* the containment guard could see it.
  Join concatenates unconditionally, leaving the guard as the decider.
- FrameServer's four empty catch blocks now record any exception not
  explained by teardown into a Fault property, asserted null by all four
  tests. The loop is detached, so previously a genuine socket/IO fault would
  have been swallowed and the test would have reported only a short served
  count with no trace of the cause.
- using var for the StringWriter; float the ratio divisor before the
  multiply rather than after.

Verified: 189/189, zero warnings; compile reaches Phase 6, exit 0, 0
findings, docs/guide/images clean.

Copilot-Session: 1bf6ef62-882e-4241-b32e-2ea22f167588

+++ tests/Reactor.DocPipeline.Tests/DocImageIntegrityTests.cs
@@ -187,7 +187,7 @@ public class DocImageIntegrityTests
                 continue;
             }
 
-            var ratio = (double)content / (region.Width * (long)region.Height);
+            var ratio = (double)content / ((double)region.Width * region.Height);
             if (ratio < minRatio)
             {
                 minRatio = ratio;
